P4CTL: Fix ovs-p4ctl to accept IP address in actions
When we pass IP address as part of action, ovs-p4ctl does not
honor this and throws an error. With new additions to p4 file,
we can expect IP address as part of actions and not just a
port number. With this change we can pass IP address as part
of actions and it will be converted to base10 value and an
gRPC message is sent to server.

Change-Id: Iab45aa412abc4457508cc6a5e160d7106a5662cb
Change-type: DefectResolution
Title: P4CTL: Fix ovs-p4ctl to accept IP address in actions
Signed-off-by: Sandeep N <sandeep.nagapattinam@intel.com>